8677q3b8q

+++ FBOLinx.Web/FBOLinx.DB/Models/SWIM/SWIMFlightLegData.cs
@@ -5,7 +5,7 @@ using System.ComponentModel.DataAnnotations.Schema;
 
 namespace FBOLinx.DB.Models
 {
-    public class SWIMFlightLegData : FBOLinxBaseEntityModel<int>
+    public class SWIMFlightLegData : FBOLinxBaseEntityModel<long>
     {
         [Required]
         public DateTime ETA { get; set; }
